small fix to Renderer

+++ FlockingBoids/Renderer/IRenderer.cs
@@ -5,7 +5,7 @@ namespace Renderer
     public interface IRenderer : IDisposable
     {
         void Clear(Color color);
-        void DrawLine(Point pt1, Point pt2, double lineWidth, Color color);
+        void DrawLine(Point pt1, Point pt2, float lineWidth, Color color);
         void FillCircle(Point center, float radius, Color color);
     }
 }
